adds nested properties merge in schemacache tests
